+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Procedure when bumping the version number:
 
 ### Added
 - `serde` feature
+- `serde::from_row_via_index`
 
 ### Changed
 - **(breaking)**